fix recursive typing issues
